pyproject.toml
[project]
name = "lvivchurch-sermons"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleh Andrus",email = "andrusoleg123@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "beautifulsoup4 (==4.12.3)",
    "box-sdk-gen (>=1.17.0,<2.0.0)",
    "gunicorn (>=21.0,<24.0)",
    "whitenoise (>=6.8,<7.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "django-two-factor-auth (>=1.18.1,<2.0.0)",
    "qrcode[pil] (>=8.2,<9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.1"
ruff = "^0.15.5"
djlint = "^1.36.4"
python-dotenv = ">=1.0"


[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["sermons_index"]

[tool.djlint]
profile = "django"
indent = 4
max_line_length = 120
